Put all data pipeline stuff into the same namespace
