+++ lib/learn_to_rank/data_pipeline/load_search_queries.rb
@@ -1,6 +1,6 @@
 require "csv"
 
-module LearnToRank
+module LearnToRank::DataPipeline
   module LoadSearchQueries
     def self.from_csv(datafile)
       queries = {}
