+++ lib/rummager.rb
@@ -102,16 +102,17 @@ require "govuk_message_queue_consumer"
 
 require "evaluate/ndcg"
 
-require "learn_to_rank/bigquery"
 require "learn_to_rank/data_pipeline"
-require "learn_to_rank/embed_features"
+require "learn_to_rank/data_pipeline/bigquery"
+require "learn_to_rank/data_pipeline/embed_features"
+require "learn_to_rank/data_pipeline/judgements_to_svm"
+require "learn_to_rank/data_pipeline/load_search_queries"
+require "learn_to_rank/data_pipeline/relevancy_judgements"
+require "learn_to_rank/data_pipeline/worker"
 require "learn_to_rank/explain_scores"
 require "learn_to_rank/feature_sets"
 require "learn_to_rank/features"
-require "learn_to_rank/judgements_to_svm"
-require "learn_to_rank/load_search_queries"
 require "learn_to_rank/ranker"
-require "learn_to_rank/relevancy_judgements"
 require "learn_to_rank/reranker"
 
 require "metasearch_index/client"
