+++ lib/tasks/learn_to_rank.rake
@@ -20,15 +20,15 @@ namespace :learn_to_rank do
   desc "Fetch data from BigQuery.  This costs money!"
   task :fetch_bigquery_export, [:credentials] do |_, args|
     assert_ltr!
-    data = LearnToRank::Bigquery.fetch(JSON.parse(Base64.decode64(args.credentials)))
+    data = LearnToRank::DataPipeline::Bigquery.fetch(JSON.parse(Base64.decode64(args.credentials)))
     export_to_csv(data, "bigquery-export")
   end
 
   desc "Export a CSV of relevancy judgements generated from CTR on popular queries"
   task :generate_relevancy_judgements, [:queries_filepath] do |_, args|
     assert_ltr!
-    queries = LearnToRank::LoadSearchQueries.from_csv(args.queries_filepath)
-    generator = LearnToRank::RelevancyJudgements.new(queries: queries)
+    queries = LearnToRank::DataPipeline::LoadSearchQueries.from_csv(args.queries_filepath)
+    generator = LearnToRank::DataPipeline::RelevancyJudgements.new(queries: queries)
     judgements = generator.relevancy_judgements
     export_to_csv(judgements, "autogenerated_relevancy_judgements")
   end
@@ -42,8 +42,8 @@ namespace :learn_to_rank do
     FileUtils.mkdir_p svm_dir
 
     judgements_data = Relevancy::LoadJudgements.from_csv(csv)
-    judgements = LearnToRank::EmbedFeatures.new(judgements_data).augmented_judgements
-    svm = LearnToRank::JudgementsToSvm.new(judgements).svm_format.group_by { |row| row.split(" ")[1] }
+    judgements = LearnToRank::DataPipeline::EmbedFeatures.new(judgements_data).augmented_judgements
+    svm = LearnToRank::DataPipeline::JudgementsToSvm.new(judgements).svm_format.group_by { |row| row.split(" ")[1] }
 
     File.open("#{svm_dir}/train.txt", "wb") do |train|
       File.open("#{svm_dir}/validate.txt", "wb") do |validate|
