+++ spec/unit/learn_to_rank/judgements_to_svm_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-RSpec.describe LearnToRank::JudgementsToSvm do
+RSpec.describe LearnToRank::DataPipeline::JudgementsToSvm do
   subject(:formatted) { described_class.new(judgements).svm_format }
 
   describe "#svm_format" do
